emi: allow setting per distro mandatory_arch
emi is using $config->{mandatory_arch} to find the list of mandatory
architectures with successful build required to upload a package.
With this change it is now also possible to set per distro
mandatory_arch in $config->{mandatory_arch}->{$distro}. If no mandatory
architecture is defined for a distro, then $config->{mandatory_arch}->{default}
is used.

git-svn-id: svn+ssh://svn.mageia.org/svn/soft/build_system/iurt/trunk@6636 5b7c3c08-40e5-403b-9995-ace06908e4af