ulri: allow setting per distro arch
ulri currently takes the list of bots defined in $config->{bot} to find
the list of architecture for which it should build packages.

With this change, it is now possible to define in $config->{$arch} the
list of architectures for each distro. If it is not defined for a
distro, $config->{$arch}->{default} is used.

git-svn-id: svn+ssh://svn.mageia.org/svn/soft/build_system/iurt/trunk@6637 5b7c3c08-40e5-403b-9995-ace06908e4af